+++ changelog.md
@@ -10,12 +10,15 @@ Started late: on 2021-04-20
   - add changelog
   - restructure main README
 
-### 2021-04-23
+### 2021-04-25
 
 - reorganize instruction set
   - add c[lear]
   - choose order of commands in menu
+- check paths for "clear" and "pdflatex" at beginning
   - choose function for command "clearpage" when session is constructed
+- reorganize settings
+
 
 ## [Summary of what happened before 2021-04-20]